refactor(finance): extract sub-components and helpers from CorrectionProposalDialog [tb-365]
Split CorrectionProposalDialogPanels.tsx (874 LOC) into 7 individual
component files under correction-proposal/:
  ContextPanel, OpsListPanel, DetailPanel, ImpactPanel,
  AiHelperPanel, BrowseRuleDetailPanel, RejectPanel

Split correction-proposal-shared.ts (245 LOC) into:
  - correction-proposal/types.ts (all types)
  - lib/normalization.ts (normalizeForMatch, transactionMatchesSignal)
  - lib/correction-utils.ts (opKindLabel, opSummary, matchTypeLabel, etc.)
  - lib/preview-scoping.ts (scopePreviewTransactions, PREVIEW_CHANGESET_MAX_TRANSACTIONS)

Original files remain as thin re-export barrels for backward compatibility.
